Basic support for FreeRTOS, still WIP

+++ modules/rtos/src/cobc/rtos/failure_handler.cpp
@@ -12,6 +12,8 @@
 #else
 #	include <cstdio>
 #	include <cstdlib>
+
+#	include <inttypes.h>
 #endif
 
 void
@@ -20,7 +22,8 @@ cobc::rtos::FailureHandler::fatal(FailureCode code)
 #ifdef __rtems__
 	rtems_fatal_error_occurred(code.getCode());
 #else
-	printf("Fatal Handler: %i\n", code.getCode());
+	//printf("Fatal Handler: %"PRIu32"\n", code.getCode());
+	printf("Fatal Handler: %i\n", static_cast<int>(code.getCode()));
 	exit(1);
 #endif
 }
